--- a/src/Csv/Csv.csproj
+++ b/src/Csv/Csv.csproj
@@ -12,6 +12,10 @@
           Visible="false" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
+  </ItemGroup>
+
   <PropertyGroup>
     <TargetFrameworks>net5.0;net6.0;net7.0;net8.0;netstandard2.1</TargetFrameworks>
     <LangVersion>12</LangVersion>