Deserialize/Serialize 100 Booleans.
```cs
[CsvObject]
public partial record Booleans
{
    [Column(0)]
    public bool Alpha { get; set; }

    [Column(1)]
    public bool Beta { get; set; }

    [Column(2)]
    public bool Gamma { get; set; }
}
```

```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.3194)
13th Gen Intel Core i7-13700F, 1 CPU, 24 logical and 16 physical cores
.NET SDK 9.0.200
  [Host]   : .NET 9.0.2 (9.0.225.6610), X64 RyuJIT AVX2
  ShortRun : .NET 9.0.2 (9.0.225.6610), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=20  LaunchCount=1
WarmupCount=10

| Method            | Mean     | Error     | StdDev    | Gen0   | Allocated |
|------------------ |---------:|----------:|----------:|-------:|----------:|
| Deserialize       | 1.187 us | 0.0056 us | 0.0064 us | 0.1068 |   1.65 KB |
| Deserialize(Fork) | 966.3 ns | 2.55 ns   | 2.94 ns   | 0.1068 |   1.65 KB |


| Method          | Mean     | Error   | StdDev  | Gen0   | Allocated |
|---------------- |---------:|--------:|--------:|-------:|----------:|
| Serialize       | 550.2 ns | 2.62 ns | 2.91 ns | 0.0401 |     632 B |
| Serialize(Fork) | 490.0 ns | 5.78 ns | 6.65 ns | 0.0401 |     632 B |
```